project 0x14

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,8 +6,8 @@
 void print_binary(unsigned long int n)
 {
 	unsigned int base = 0;
-	unsigned int sum = 0;
-	unsigned int value;
+	unsigned long int sum = 0;
+	unsigned long int value;
 
 	if (n == 0)
 	{
@@ -28,6 +28,7 @@ void print_binary(unsigned long int n)
 	while (n > 0 || base > 0)
 	{
 		value = _pow(2, base - 1);
+
 		if (value <= n)
 		{
 			_putchar('1');
@@ -53,6 +54,7 @@ void print_binary(unsigned long int n)
 int _pow(int b, int exp)
 {
 	int result = 1;
+
 	if (exp == 0)
 		return (1);
 	while (exp > 0)
